Refacored RedisCache service to use Strings and expire functionality
Each delta is stored in as a Redis string value with it's own key.
An expiry time is set on the key which Redis automatically handles so no need for manual delete.
The offsets are stored in a stoerted set with the oddset as key and the corresponding delta key at its value.